Agent_Based_Samples/market_research_analyst/requirements.txt
redis==5.0.4
docker==7.1.0
aiohttp==3.11.14
python-dotenv==1.0.1
aiohttp_cors==0.7.0
setuptools==75.8.0
openai==1.86.0
azure-identity==1.20.0
azure-storage-blob==12.19.1
azure-keyvault-secrets==4.7.0
azure-ai-projects==1.0.0b11
semantic-kernel[azure]==1.33.0
opentelemetry-api==1.29.0
opentelemetry-exporter-otlp==1.29.0
azure-monitor-opentelemetry==1.6.4
opencensus==0.11.2
opencensus-ext-azure==1.1.9
azure-search-documents==11.6.0b4